crawler/requirements.txt
pandas
requests
lxml
python-dotenv
supabase
